refactor : #143 공통 서비스 코드 (Kr, Eng 통신 코드) 이전
