+++ src/main/java/com/mingles/metamingle/interactivemovie/command/application/service/InteractiveMovieCommandService.java
@@ -71,55 +71,12 @@ public class InteractiveMovieCommandService {
         this.webClient = WebClient.builder().baseUrl(aiAddr).build();
     }
 
-
     private final InteractiveMovieCommandRepository interactiveMovieCommandRepository;
     private final InteractiveMovieDomainService interactiveMovieDomainService;
     private final ApiShortFormService apiShortFormService;
     private final QuizCommandService quizCommandService;
     private final AiInfraService aiInfraService;
 
-//    public List<CreateInteractiveMovieResponse> createInteractiveMovie(List<MultipartFile> files, String title, String description, List<String> choices, Long memberNo)
-//            throws JCodecException, IOException {
-//
-//        List<CreateInteractiveMovieResponse> response = new ArrayList<>();
-//
-//        ShortForm shortForm = apiShortFormService.createShortFormWithInteractiveMovie(files.get(0), title, description, memberNo);
-//        response.add(new CreateInteractiveMovieResponse(shortForm.getShortFormNo(), null, shortForm.getThumbnailUrl(),shortForm.getUrl(), "none", 0));
-//
-//        ShortFormNoVO shortFormNoVO = new ShortFormNoVO(shortForm.getShortFormNo());
-//
-//        for (int i = 1; i <= 2; i++) {
-//
-//            String fileKeyName = createFileName(files.get(i).getOriginalFilename()); // 파일 이름을 고유한 파일 이름으로 교체
-//
-//            Bucket bucket = StorageClient.getInstance().bucket(bucketName);
-//            InputStream inputStream = files.get(i).getInputStream();
-//            Blob blob = bucket.create(fileKeyName, inputStream, files.get(i).getContentType());
-//
-//            inputStream.close();
-//
-//            String url = bucketUrl + fileKeyName + "?alt=media";
-//
-//            String thumbnailUrl = createAndUploadThumbnail(files.get(i), fileKeyName);
-//
-//            // 임시 멤버 넘버 사용
-//            MemberNoVO memberNoVO = new MemberNoVO(memberNo);
-//
-//            InteractiveMovie interactiveMovieEntity = new InteractiveMovie(title, url, thumbnailUrl, description, choices.get(i - 1), new Date(),
-//                    i, shortFormNoVO, memberNoVO);
-//
-//            InteractiveMovie uploadedInteractiveMovie = interactiveMovieCommandRepository.save(interactiveMovieEntity);
-//
-//            response.add(new CreateInteractiveMovieResponse(null, uploadedInteractiveMovie.getInteractiveMovieNo(),
-//                                                            uploadedInteractiveMovie.getThumbnailUrl(),
-//                                                            uploadedInteractiveMovie.getUrl(),
-//                                                            uploadedInteractiveMovie.getChoice(),
-//                                                            uploadedInteractiveMovie.getSequence()));
-//        }
-//
-//        return response;
-//
-//    }
     @Async
     @Transactional
     public List<CreateInteractiveMovieResponse> createInteractiveMovieWithSubtitle(List<byte[]> files, List<String> fileNames, String uuid,
@@ -167,10 +124,10 @@ public class InteractiveMovieCommandService {
         SubtitledVideo subtitledVideo = new SubtitledVideo();
 
         System.out.println("ai 영어 자막 영상 요청 : 인터랙티브 무비");
-        subtitledVideo.setFileEng(sendToAIForEngSub(file.getResource(), fileKeyName));
+        subtitledVideo.setFileEng(aiInfraService.sendToAIForEngSub(file.getResource(), fileKeyName));
         System.out.println("ai 영어 자막 영상 응답 완료 : 인터랙티브 무비");
         System.out.println("ai 한글 자막 영상 요청 : 인터랙티브 무비");
-        subtitledVideo.setFileKr(sendToAIForKrSub(fileKeyName));
+        subtitledVideo.setFileKr(aiInfraService.sendToAIForKrSub(fileKeyName));
         System.out.println("ai 한글 자막 영상 응답 완료 : 인터랙티브 무비");
 
         // 썸네일 이미지 생성 (영어자막/한글자막 영상에 관련없이 썸네일을 같음)
@@ -195,44 +152,6 @@ public class InteractiveMovieCommandService {
                 createdMovie.getChoice(),
                 createdMovie.getSequence());
     }
-
-    private MultipartFile sendToAIForEngSub(Resource file, String fileKeyName) {
-
-        MultipartBodyBuilder bodyBuilder = new MultipartBodyBuilder();
-        bodyBuilder.part("file", file);
-        bodyBuilder.part("file_uuid", fileKeyName);
-
-        System.out.println("영어 자막 영상 처리 중");
-
-        Flux<DataBuffer> responseBody = webClient.post()
-                .uri("/mp4/en_script_video/")
-                .contentType(MediaType.MULTIPART_FORM_DATA)
-                .body(BodyInserters.fromMultipartData(bodyBuilder.build()))
-                .accept(MediaType.APPLICATION_OCTET_STREAM)
-                .retrieve()
-                .bodyToFlux(DataBuffer.class);
-
-        return dataBufferToMultipartFile(fileKeyName, responseBody);
-    }
-
-    private MultipartFile sendToAIForKrSub(String fileKeyName) {
-
-        MultipartBodyBuilder bodyBuilder = new MultipartBodyBuilder();
-        bodyBuilder.part("file_uuid", fileKeyName);
-
-        System.out.println("한글 자막 영상 처리 중");
-
-        Flux<DataBuffer> responseBody = webClient.post()
-                .uri("/mp4/modi_kr_script_video/")
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(BodyInserters.fromMultipartData(bodyBuilder.build()))
-                .accept(MediaType.APPLICATION_OCTET_STREAM)
-                .retrieve()
-                .bodyToFlux(DataBuffer.class);
-
-        return dataBufferToMultipartFile(fileKeyName, responseBody);
-    }
-
 
     private UploadVideo createInteractiveMovie(MultipartFile file, String fileKeyName, String thumbnailUrl)
                                                     throws IOException {
